km_system/requirements.txt
Django==1.7
Fabric==1.10.0
argparse==1.2.1
django-braces==1.4.0
django-pagination==1.0.5
ecdsa==0.11
paramiko==1.15.1
pycrypto==2.6.1
six==1.8.0
wsgiref==0.1.2
